Improvements in the timing
Compile the vector add function for best performance.

Replace the timeit runner with a loop so the synchronization time
can be included.

The calibration loop includes a synchronization after each function
call.  This makes it easier to call the timeit autorange, at the
cost of some accuracy.  It should still be adequate for running
the benchmark loop.

The calibration loop could be made more accurate by calling the
synchronization after an increasing number of inner loops.

Flush the output after each line so results can be viewed immediately if
the output is redirected or piped.  I typically pipe the output to
'tee' so it can be seen and saved to a file.